+++ websieve/profiles/register.js
@@ -5,6 +5,7 @@
   registerGlassdoorProfiles();
   registerRingierAxelSpringerProfiles();
   registerMediaCommerceLeadProfiles();
+  registerHostAwareProfile(true, softwareProjectHomepageContent);
   registerNewsHomepageProfiles();
   registerWpHomepageProfile();
   registerOnetHomepageProfile();